Simple way to obtain data mining rules

This is a rule-based machine learning method to discover interesting relationships between a consequent and an antecedent (or group of antecedents) in large databases.

AsociationRules(data, Support = 0.2, Confidence = 0.1, minlength = 2)

Arguments

data
a data frame with discrete variables.

Support
a numeric value for the minimun support of the antecedents (default: 0.2).

Confidence
a numeric value for the minumun confidence of confidence in rule/association method (default: 0.8)

minlength
an integer value for the minimal number of items per item set (default: 2 item)

Value

A MLA object of subclass AsociationRules

Examples

## Load a Dataset
data(EGATUR)
## Generate an asociation rules with apriori, remmember only support discretized variables,
##  in this remove numerical variables.
Rules <- AsociationRules(EGATUR[,c(2,4,5,8)])

#> Apriori
#> 
#> Parameter specification:
#>  confidence minval smax arem  aval originalSupport maxtime support minlen
#>         0.1    0.1    1 none FALSE            TRUE       5     0.2      2
#>  maxlen target   ext
#>      10  rules FALSE
#> 
#> Algorithmic control:
#>  filter tree heap memopt load sort verbose
#>     0.1 TRUE TRUE  FALSE TRUE    2    TRUE
#> 
#> Absolute minimum support count: 3054 
#> 
#> set item appearances ...[0 item(s)] done [0.00s].
#> set transactions ...[24 item(s), 15270 transaction(s)] done [0.01s].
#> sorting and recoding items ... [4 item(s)] done [0.00s].
#> creating transaction tree ... done [0.01s].
#> checking subsets of size 1 2 3 done [0.00s].
#> writing ... [11 rule(s)] done [0.00s].
#> creating S4 object  ... done [0.00s].