An expert rule based system using Reduction Based on Significance

CREA-RBS is a rule reduction method for allocating a significance value to each rule in the system so that experts may select the rules that should be considered as preferable and understand the exact degree of correlation between the different rule attributes.

Arguments

formula
a formula of the form y ~ x1 + x2 + ...

data
the data frame that contains the variables specified in formula.

Value

A MLA object of subclass CREA-RBS

Details

Significance is calculated from the antecedent frequency and rule frequency parameters for each rule; if the first one is above the minimal level and rule frequency is in a critical interval, its significance ratio is computed by the algorithm. These critical boundaries are calculated by an incremental method and the rule space is divided according to them. The significance function is defined for these intervals.

References

Almiñana, M., Escudero, L. F., Pérez-Martín, A., Rabasa, A., & Santamaría, L. (2014). A classification rule reduction algorithm based on significance domains. Top, 22(1), 397-418.

Examples

## Load a Dataset
data(EGATUR)
## Generate a CREA-RBS model, remmember only support discretized variables
CREA.RBS(GastoTotalD~pais+aloja+motivo,data=EGATUR)

#>                 pais                          aloja   motivo
#> 1                 UK                         Hotels  Leisure
#> 2            Germany                         Hotels  Leisure
#> 3             France                         Hotels  Leisure
#> 4                 UK Over-The-Counter Accommodation  Leisure
#> 5              Italy                         Hotels  Leisure
#> 6             France Over-The-Counter Accommodation  Leisure
#> 7   Nordic countries                         Hotels  Leisure
#> 8                 UK Over-The-Counter Accommodation   Others
#> 9        Netherlands                         Hotels  Leisure
#> 10                UK                         Hotels  Leisure
#> 11                UK Over-The-Counter Accommodation  Leisure
#> 12       Netherlands                         Hotels  Leisure
#> 13            France Over-The-Counter Accommodation  Leisure
#> 14            France                         Hotels  Leisure
#> 15             Italy                         Hotels  Leisure
#> 16                UK Over-The-Counter Accommodation   Others
#> 17     Rest of world                         Hotels  Leisure
#> 18   Rest of America                         Hotels  Leisure
#> 19  Nordic countries                         Hotels  Leisure
#> 20                UK                         Hotels  Leisure
#> 21                UK Over-The-Counter Accommodation  Leisure
#> 22                UK Over-The-Counter Accommodation   Others
#> 23          Portugal Over-The-Counter Accommodation   Others
#> 24       Switzerland                         Hotels Business
#> 25          Portugal                         Hotels   Others
#> 26             Italy                         Hotels   Others
#> 27          Portugal Over-The-Counter Accommodation  Leisure
#> 28                UK                 Rest of market   Others
#> 29           Germany Over-The-Counter Accommodation Business
#> 30            France Over-The-Counter Accommodation Business
#> 31       Switzerland                         Hotels   Others
#> 32   Rest of America                         Hotels   Others
#> 33             Rusia Over-The-Counter Accommodation  Leisure
#> 34           Germany                 Rest of market   Others
#> 35             Italy Over-The-Counter Accommodation Business
#> 36       Netherlands                         Hotels   Others
#> 37           Ireland                         Hotels Business
#> 38  Nordic countries                         Hotels   Others
#> 39       Switzerland Over-The-Counter Accommodation Business
#> 40               USA                 Rest of market  Leisure
#> 41               USA                 Rest of market  Leisure
#> 42            France                 Rest of market   Others
#> 43          Portugal Over-The-Counter Accommodation Business
#> 44           Belgium                         Hotels   Others
#> 45             Italy                 Rest of market   Others
#> 46    Rest of Europe Over-The-Counter Accommodation Business
#> 47   Rest of America                         Hotels   Others
#> 48   Rest of America                 Rest of market   Others
#> 49    Rest of Europe                         Hotels   Others
#> 50               USA                         Hotels   Others
#> 51               USA                         Hotels   Others
#> 52           Belgium Over-The-Counter Accommodation Business
#> 53            France                 Rest of market Business
#> 54             Italy                 Rest of market Business
#> 55   Rest of America                 Rest of market   Others
#> 56    Rest of Europe                 Rest of market Business
#> 57             Rusia Over-The-Counter Accommodation   Others
#> 58       Netherlands Over-The-Counter Accommodation Business
#> 59   Rest of America Over-The-Counter Accommodation Business
#> 60          Portugal                 Rest of market  Leisure
#> 61             Rusia                         Hotels Business
#> 62  Nordic countries Over-The-Counter Accommodation Business
#> 63  Nordic countries                 Rest of market   Others
#> 64                UK                 Rest of market   Others
#> 65           Ireland                         Hotels   Others
#> 66     Rest of world                 Rest of market Business
#> 67             Rusia                 Rest of market  Leisure
#> 68             Rusia                 Rest of market  Leisure
#> 69                UK                 Rest of market Business
#> 70           Germany                 Rest of market Business
#> 71           Germany                 Rest of market   Others
#> 72  Nordic countries                 Rest of market Business
#> 73   Rest of America                 Rest of market Business
#> 74    Rest of Europe Over-The-Counter Accommodation Business
#> 75     Rest of world                 Rest of market   Others
#> 76            France                 Rest of market   Others
#> 77             Italy                 Rest of market   Others
#> 78             Italy                 Rest of market   Others
#> 79       Netherlands                 Rest of market   Others
#> 80    Rest of Europe                 Rest of market Business
#> 81     Rest of world                 Rest of market Business
#> 82             Rusia Over-The-Counter Accommodation  Leisure
#> 83               USA                 Rest of market   Others
#> 84            France                 Rest of market   Others
#> 85           Ireland                 Rest of market   Others
#> 86       Netherlands                 Rest of market Business
#> 87   Rest of America Over-The-Counter Accommodation Business
#> 88    Rest of Europe                 Rest of market   Others
#> 89     Rest of world Over-The-Counter Accommodation Business
#> 90     Rest of world Over-The-Counter Accommodation Business
#> 91     Rest of world                 Rest of market   Others
#> 92             Rusia                         Hotels Business
#> 93             Rusia                         Hotels   Others
#> 94             Rusia Over-The-Counter Accommodation   Others
#> 95       Switzerland                 Rest of market   Others
#> 96                UK                 Rest of market Business
#> 97           Belgium                 Rest of market Business
#> 98           Belgium                 Rest of market   Others
#> 99           Germany Over-The-Counter Accommodation Business
#> 100          Germany                 Rest of market   Others
#> 101            Italy Over-The-Counter Accommodation Business
#> 102            Italy                 Rest of market Business
#> 103         Portugal                 Rest of market Business
#> 104         Portugal                 Rest of market   Others
#> 105   Rest of Europe                         Hotels   Others
#> 106   Rest of Europe                 Rest of market   Others
#> 107    Rest of world                 Rest of market Business
#> 108      Switzerland                 Rest of market   Others
#> 109              USA Over-The-Counter Accommodation Business
#> 110              USA Over-The-Counter Accommodation Business
#> 111              USA                 Rest of market   Others
#> 112          Belgium Over-The-Counter Accommodation Business
#> 113          Belgium                 Rest of market   Others
#> 114          Belgium                 Rest of market   Others
#> 115          Germany                 Rest of market Business
#> 116          Ireland Over-The-Counter Accommodation Business
#> 117          Ireland                 Rest of market Business
#> 118      Netherlands                         Hotels   Others
#> 119      Netherlands                 Rest of market Business
#> 120      Netherlands                 Rest of market Business
#> 121      Netherlands                 Rest of market   Others
#> 122      Netherlands                 Rest of market   Others
#> 123 Nordic countries Over-The-Counter Accommodation Business
#> 124 Nordic countries                 Rest of market   Others
#> 125 Nordic countries                 Rest of market   Others
#> 126  Rest of America                         Hotels   Others
#> 127  Rest of America                 Rest of market Business
#> 128  Rest of America                 Rest of market   Others
#> 129   Rest of Europe                 Rest of market   Others
#> 130    Rest of world Over-The-Counter Accommodation Business
#> 131    Rest of world                 Rest of market   Others
#> 132            Rusia Over-The-Counter Accommodation Business
#> 133      Switzerland                         Hotels Business
#> 134      Switzerland Over-The-Counter Accommodation Business
#> 135      Switzerland                 Rest of market Business
#> 136              USA                 Rest of market Business
#> 137              USA                 Rest of market Business
#> 138              USA                 Rest of market   Others
#>             GastoTotalD totalRegla total.casos      Support   Confidence Region
#> 1         [17,1.67e+03)       1407        1471 9.633268e-02 0.9564921822      2
#> 2         [17,1.67e+03)        884        1007 6.594630e-02 0.8778550149      2
#> 3         [17,1.67e+03)        796         809 5.297970e-02 0.9839307787      2
#> 4         [17,1.67e+03)        616         680 4.453176e-02 0.9058823529      2
#> 5         [17,1.67e+03)        484         494 3.235102e-02 0.9797570850      2
#> 6         [17,1.67e+03)        469         488 3.195809e-02 0.9610655738      2
#> 7         [17,1.67e+03)        455         556 3.641126e-02 0.8183453237      2
#> 8         [17,1.67e+03)        363         373 2.442698e-02 0.9731903485      2
#> 9         [17,1.67e+03)        337         363 2.377210e-02 0.9283746556      2
#> 10  [1.67e+03,5.51e+03)         63        1471 9.633268e-02 0.0428280082      1
#> 11  [1.67e+03,5.51e+03)         63         680 4.453176e-02 0.0926470588      1
#> 12  [1.67e+03,5.51e+03)         26         363 2.377210e-02 0.0716253444      1
#> 13  [1.67e+03,5.51e+03)         19         488 3.195809e-02 0.0389344262      1
#> 14  [1.67e+03,5.51e+03)         13         809 5.297970e-02 0.0160692213      1
#> 15  [1.67e+03,5.51e+03)         10         494 3.235102e-02 0.0202429150      1
#> 16  [1.67e+03,5.51e+03)          9         373 2.442698e-02 0.0241286863      1
#> 17  [5.51e+03,1.99e+04]          7         477 3.123772e-02 0.0146750524      1
#> 18  [5.51e+03,1.99e+04]          3         378 2.475442e-02 0.0079365079      1
#> 19  [5.51e+03,1.99e+04]          2         556 3.641126e-02 0.0035971223      1
#> 20  [5.51e+03,1.99e+04]          1        1471 9.633268e-02 0.0006798097      1
#> 21  [5.51e+03,1.99e+04]          1         680 4.453176e-02 0.0014705882      1
#> 22  [5.51e+03,1.99e+04]          1         373 2.442698e-02 0.0026809651      1
#> 23        [17,1.67e+03)         34          34 2.226588e-03 1.0000000000      3
#> 24        [17,1.67e+03)         31          32 2.095612e-03 0.9687500000      3
#> 25        [17,1.67e+03)         30          30 1.964637e-03 1.0000000000      3
#> 26        [17,1.67e+03)         29          29 1.899149e-03 1.0000000000      3
#> 27        [17,1.67e+03)         26          26 1.702685e-03 1.0000000000      3
#> 28        [17,1.67e+03)         25          32 2.095612e-03 0.7812500000      3
#> 29        [17,1.67e+03)         24          26 1.702685e-03 0.9230769231      3
#> 30        [17,1.67e+03)         23          23 1.506221e-03 1.0000000000      3
#> 31        [17,1.67e+03)         23          23 1.506221e-03 1.0000000000      3
#> 32  [1.67e+03,5.51e+03)         22          35 2.292076e-03 0.6285714286      3
#> 33        [17,1.67e+03)         22          26 1.702685e-03 0.8461538462      3
#> 34        [17,1.67e+03)         20          27 1.768173e-03 0.7407407407      3
#> 35        [17,1.67e+03)         18          20 1.309758e-03 0.9000000000      3
#> 36        [17,1.67e+03)         18          19 1.244270e-03 0.9473684211      3
#> 37        [17,1.67e+03)         17          17 1.113294e-03 1.0000000000      3
#> 38        [17,1.67e+03)         17          17 1.113294e-03 1.0000000000      3
#> 39        [17,1.67e+03)         17          18 1.178782e-03 0.9444444444      3
#> 40        [17,1.67e+03)         16          31 2.030124e-03 0.5161290323      3
#> 41  [1.67e+03,5.51e+03)         15          31 2.030124e-03 0.4838709677      3
#> 42        [17,1.67e+03)         14          21 1.375246e-03 0.6666666667      3
#> 43        [17,1.67e+03)         14          14 9.168304e-04 1.0000000000      3
#> 44        [17,1.67e+03)         13          13 8.513425e-04 1.0000000000      3
#> 45        [17,1.67e+03)         13          21 1.375246e-03 0.6190476190      3
#> 46        [17,1.67e+03)         13          18 1.178782e-03 0.7222222222      3
#> 47        [17,1.67e+03)         12          35 2.292076e-03 0.3428571429      3
#> 48  [5.51e+03,1.99e+04]         12          23 1.506221e-03 0.5217391304      3
#> 49        [17,1.67e+03)         12          14 9.168304e-04 0.8571428571      3
#> 50        [17,1.67e+03)         11          22 1.440733e-03 0.5000000000      3
#> 51  [1.67e+03,5.51e+03)         11          22 1.440733e-03 0.5000000000      3
#> 52        [17,1.67e+03)         10          11 7.203667e-04 0.9090909091      3
#> 53        [17,1.67e+03)         10          10 6.548788e-04 1.0000000000      3
#> 54        [17,1.67e+03)         10          12 7.858546e-04 0.8333333333      3
#> 55  [1.67e+03,5.51e+03)         10          23 1.506221e-03 0.4347826087      3
#> 56        [17,1.67e+03)         10          14 9.168304e-04 0.7142857143      3
#> 57        [17,1.67e+03)         10          13 8.513425e-04 0.7692307692      3
#> 58        [17,1.67e+03)          9           9 5.893910e-04 1.0000000000      3
#> 59  [1.67e+03,5.51e+03)          9          12 7.858546e-04 0.7500000000      3
#> 60        [17,1.67e+03)          8           8 5.239031e-04 1.0000000000      3
#> 61        [17,1.67e+03)          8          11 7.203667e-04 0.7272727273      3
#> 62        [17,1.67e+03)          7           8 5.239031e-04 0.8750000000      3
#> 63  [1.67e+03,5.51e+03)          7           9 5.893910e-04 0.7777777778      3
#> 64  [1.67e+03,5.51e+03)          7          32 2.095612e-03 0.2187500000      3
#> 65        [17,1.67e+03)          6           6 3.929273e-04 1.0000000000      3
#> 66  [1.67e+03,5.51e+03)          6          12 7.858546e-04 0.5000000000      3
#> 67        [17,1.67e+03)          6          12 7.858546e-04 0.5000000000      3
#> 68  [1.67e+03,5.51e+03)          6          12 7.858546e-04 0.5000000000      3
#> 69        [17,1.67e+03)          6           9 5.893910e-04 0.6666666667      3
#> 70        [17,1.67e+03)          5           6 3.929273e-04 0.8333333333      3
#> 71  [1.67e+03,5.51e+03)          5          27 1.768173e-03 0.1851851852      3
#> 72        [17,1.67e+03)          5           5 3.274394e-04 1.0000000000      3
#> 73  [5.51e+03,1.99e+04]          5           6 3.929273e-04 0.8333333333      3
#> 74  [1.67e+03,5.51e+03)          5          18 1.178782e-03 0.2777777778      3
#> 75  [5.51e+03,1.99e+04]          5           9 5.893910e-04 0.5555555556      3
#> 76  [1.67e+03,5.51e+03)          4          21 1.375246e-03 0.1904761905      3
#> 77  [1.67e+03,5.51e+03)          4          21 1.375246e-03 0.1904761905      3
#> 78  [5.51e+03,1.99e+04]          4          21 1.375246e-03 0.1904761905      3
#> 79        [17,1.67e+03)          4           6 3.929273e-04 0.6666666667      3
#> 80  [1.67e+03,5.51e+03)          4          14 9.168304e-04 0.2857142857      3
#> 81        [17,1.67e+03)          4          12 7.858546e-04 0.3333333333      3
#> 82  [1.67e+03,5.51e+03)          4          26 1.702685e-03 0.1538461538      3
#> 83  [5.51e+03,1.99e+04]          4           7 4.584152e-04 0.5714285714      3
#> 84  [5.51e+03,1.99e+04]          3          21 1.375246e-03 0.1428571429      3
#> 85        [17,1.67e+03)          3           3 1.964637e-04 1.0000000000      3
#> 86        [17,1.67e+03)          3           5 3.274394e-04 0.6000000000      3
#> 87        [17,1.67e+03)          3          12 7.858546e-04 0.2500000000      3
#> 88        [17,1.67e+03)          3           6 3.929273e-04 0.5000000000      3
#> 89        [17,1.67e+03)          3           7 4.584152e-04 0.4285714286      3
#> 90  [1.67e+03,5.51e+03)          3           7 4.584152e-04 0.4285714286      3
#> 91  [1.67e+03,5.51e+03)          3           9 5.893910e-04 0.3333333333      3
#> 92  [1.67e+03,5.51e+03)          3          11 7.203667e-04 0.2727272727      3
#> 93        [17,1.67e+03)          3           3 1.964637e-04 1.0000000000      3
#> 94  [1.67e+03,5.51e+03)          3          13 8.513425e-04 0.2307692308      3
#> 95        [17,1.67e+03)          3           5 3.274394e-04 0.6000000000      3
#> 96  [1.67e+03,5.51e+03)          3           9 5.893910e-04 0.3333333333      3
#> 97        [17,1.67e+03)          2           2 1.309758e-04 1.0000000000      3
#> 98  [5.51e+03,1.99e+04]          2           4 2.619515e-04 0.5000000000      3
#> 99  [1.67e+03,5.51e+03)          2          26 1.702685e-03 0.0769230769      3
#> 100 [5.51e+03,1.99e+04]          2          27 1.768173e-03 0.0740740741      3
#> 101 [1.67e+03,5.51e+03)          2          20 1.309758e-03 0.1000000000      3
#> 102 [1.67e+03,5.51e+03)          2          12 7.858546e-04 0.1666666667      3
#> 103       [17,1.67e+03)          2           2 1.309758e-04 1.0000000000      3
#> 104       [17,1.67e+03)          2           2 1.309758e-04 1.0000000000      3
#> 105 [1.67e+03,5.51e+03)          2          14 9.168304e-04 0.1428571429      3
#> 106 [5.51e+03,1.99e+04]          2           6 3.929273e-04 0.3333333333      3
#> 107 [5.51e+03,1.99e+04]          2          12 7.858546e-04 0.1666666667      3
#> 108 [1.67e+03,5.51e+03)          2           5 3.274394e-04 0.4000000000      3
#> 109       [17,1.67e+03)          2           4 2.619515e-04 0.5000000000      3
#> 110 [1.67e+03,5.51e+03)          2           4 2.619515e-04 0.5000000000      3
#> 111 [1.67e+03,5.51e+03)          2           7 4.584152e-04 0.2857142857      3
#> 112 [1.67e+03,5.51e+03)          1          11 7.203667e-04 0.0909090909      3
#> 113       [17,1.67e+03)          1           4 2.619515e-04 0.2500000000      3
#> 114 [1.67e+03,5.51e+03)          1           4 2.619515e-04 0.2500000000      3
#> 115 [1.67e+03,5.51e+03)          1           6 3.929273e-04 0.1666666667      3
#> 116       [17,1.67e+03)          1           1 6.548788e-05 1.0000000000      3
#> 117       [17,1.67e+03)          1           1 6.548788e-05 1.0000000000      3
#> 118 [1.67e+03,5.51e+03)          1          19 1.244270e-03 0.0526315789      3
#> 119 [1.67e+03,5.51e+03)          1           5 3.274394e-04 0.2000000000      3
#> 120 [5.51e+03,1.99e+04]          1           5 3.274394e-04 0.2000000000      3
#> 121 [1.67e+03,5.51e+03)          1           6 3.929273e-04 0.1666666667      3
#> 122 [5.51e+03,1.99e+04]          1           6 3.929273e-04 0.1666666667      3
#> 123 [1.67e+03,5.51e+03)          1           8 5.239031e-04 0.1250000000      3
#> 124       [17,1.67e+03)          1           9 5.893910e-04 0.1111111111      3
#> 125 [5.51e+03,1.99e+04]          1           9 5.893910e-04 0.1111111111      3
#> 126 [5.51e+03,1.99e+04]          1          35 2.292076e-03 0.0285714286      3
#> 127 [1.67e+03,5.51e+03)          1           6 3.929273e-04 0.1666666667      3
#> 128       [17,1.67e+03)          1          23 1.506221e-03 0.0434782609      3
#> 129 [1.67e+03,5.51e+03)          1           6 3.929273e-04 0.1666666667      3
#> 130 [5.51e+03,1.99e+04]          1           7 4.584152e-04 0.1428571429      3
#> 131       [17,1.67e+03)          1           9 5.893910e-04 0.1111111111      3
#> 132       [17,1.67e+03)          1           1 6.548788e-05 1.0000000000      3
#> 133 [1.67e+03,5.51e+03)          1          32 2.095612e-03 0.0312500000      3
#> 134 [1.67e+03,5.51e+03)          1          18 1.178782e-03 0.0555555556      3
#> 135 [1.67e+03,5.51e+03)          1           1 6.548788e-05 1.0000000000      3
#> 136 [1.67e+03,5.51e+03)          1           2 1.309758e-04 0.5000000000      3
#> 137 [5.51e+03,1.99e+04]          1           2 1.309758e-04 0.5000000000      3
#> 138       [17,1.67e+03)          1           7 4.584152e-04 0.1428571429      3
#>      Importance
#> 1    0.09214145
#> 2    0.05789129
#> 3    0.05212836
#> 4    0.04034054
#> 5    0.03169614
#> 6    0.03071382
#> 7    0.02979699
#> 8    0.02377210
#> 9    0.02206942
#> 10  -0.99587426
#> 11  -0.99587426
#> 12  -0.99829731
#> 13  -0.99875573
#> 14  -0.99914866
#> 15  -0.99934512
#> 16  -0.99941061
#> 17  -0.99954158
#> 18  -0.99980354
#> 19  -0.99986902
#> 20  -0.99993451
#> 21  -0.99993451
#> 22  -0.99993451
#> 23  -1.99777341
#> 24  -1.99796988
#> 25  -1.99803536
#> 26  -1.99810085
#> 27  -1.99829731
#> 28  -1.99836280
#> 29  -1.99842829
#> 30  -1.99849378
#> 31  -1.99849378
#> 32  -1.99855927
#> 33  -1.99855927
#> 34  -1.99869024
#> 35  -1.99882122
#> 36  -1.99882122
#> 37  -1.99888671
#> 38  -1.99888671
#> 39  -1.99888671
#> 40  -1.99895219
#> 41  -1.99901768
#> 42  -1.99908317
#> 43  -1.99908317
#> 44  -1.99914866
#> 45  -1.99914866
#> 46  -1.99914866
#> 47  -1.99921415
#> 48  -1.99921415
#> 49  -1.99921415
#> 50  -1.99927963
#> 51  -1.99927963
#> 52  -1.99934512
#> 53  -1.99934512
#> 54  -1.99934512
#> 55  -1.99934512
#> 56  -1.99934512
#> 57  -1.99934512
#> 58  -1.99941061
#> 59  -1.99941061
#> 60  -1.99947610
#> 61  -1.99947610
#> 62  -1.99954158
#> 63  -1.99954158
#> 64  -1.99954158
#> 65  -1.99960707
#> 66  -1.99960707
#> 67  -1.99960707
#> 68  -1.99960707
#> 69  -1.99960707
#> 70  -1.99967256
#> 71  -1.99967256
#> 72  -1.99967256
#> 73  -1.99967256
#> 74  -1.99967256
#> 75  -1.99967256
#> 76  -1.99973805
#> 77  -1.99973805
#> 78  -1.99973805
#> 79  -1.99973805
#> 80  -1.99973805
#> 81  -1.99973805
#> 82  -1.99973805
#> 83  -1.99973805
#> 84  -1.99980354
#> 85  -1.99980354
#> 86  -1.99980354
#> 87  -1.99980354
#> 88  -1.99980354
#> 89  -1.99980354
#> 90  -1.99980354
#> 91  -1.99980354
#> 92  -1.99980354
#> 93  -1.99980354
#> 94  -1.99980354
#> 95  -1.99980354
#> 96  -1.99980354
#> 97  -1.99986902
#> 98  -1.99986902
#> 99  -1.99986902
#> 100 -1.99986902
#> 101 -1.99986902
#> 102 -1.99986902
#> 103 -1.99986902
#> 104 -1.99986902
#> 105 -1.99986902
#> 106 -1.99986902
#> 107 -1.99986902
#> 108 -1.99986902
#> 109 -1.99986902
#> 110 -1.99986902
#> 111 -1.99986902
#> 112 -1.99993451
#> 113 -1.99993451
#> 114 -1.99993451
#> 115 -1.99993451
#> 116 -1.99993451
#> 117 -1.99993451
#> 118 -1.99993451
#> 119 -1.99993451
#> 120 -1.99993451
#> 121 -1.99993451
#> 122 -1.99993451
#> 123 -1.99993451
#> 124 -1.99993451
#> 125 -1.99993451
#> 126 -1.99993451
#> 127 -1.99993451
#> 128 -1.99993451
#> 129 -1.99993451
#> 130 -1.99993451
#> 131 -1.99993451
#> 132 -1.99993451
#> 133 -1.99993451
#> 134 -1.99993451
#> 135 -1.99993451
#> 136 -1.99993451
#> 137 -1.99993451
#> 138 -1.99993451