EGATUR dataset

Tourist Expenditure Survey (EGATUR) is the response by the Spanish Tourist Authorities to the growing need for information by this sector.

data("EGATUR")

Format

A data frame with 30541 observations on the following 13 variables.

mm_aaaa

a numeric vector

pais

a factor with country names

A13

a numeric vector

aloja

a factor with levels Hotels Rest of market Over-The-Counter Accommodation

motivo

a factor with levels Leisure Business Others

gastototal

a numeric vector

factoregatur

a numeric vector

GastoTotalD

a factor with levels [17,1.67e+03) [1.67e+03,5.51e+03) [5.51e+03,1.99e+04]

A13_D

a factor with levels [1,3) [3,4) [4,6) [6,7) [7,14) [14,180]

Details

Tourist Expenditure Survey (EGATUR) is the response by the Spanish Tourist Authorities to the growing need for information by this sector, which is one of the major driving forces of the Spanish economy. The information provided by this survey makes it possible to ascertain with a greater degree of precision the volume of tourist expenditure by foreign visitors coming to Spain each month by different concepts, and to also analyse key aspects of their tourist behaviour. EGATUR makes it possible to improve strategic knowledge of variables regarding fundamental expenditure and tourist behaviour by visitors from other countries, and to a large extent compensate the loss of information which, in order to estimate the income and payment entries for tourism in the Balance of Payments, was being used by the Bank of Spain prior to the introduction of the Euro. Likewise it provides highly relevant information in terms of National Accounts estimates and, in particular, in estimating the main groups of the recent Tourism Satellite Account for Spain.

References

‘EGATUR’, available at http://www.ine.es/dyngs/inebase/es/operacion.htm?c=estadistica_c&cid=1254736177002&menu=ultidatos&idp=1254735576863

Examples

data(EGATUR)